Added maps

+++ app/src/main/java/com/shuvam/recsnd/MyLocation.java
@@ -10,23 +10,23 @@ public class MyLocation {
 
 
     @SerializedName("lat")
-    Long latitude;
+    Double latitude;
     @SerializedName("lng")
-    Long longitude;
+    Double longitude;
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
